requirements.txt
click==7.1.1
cssselect==1.1.0
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
lxml==4.5.0
MarkupSafe==1.1.1
parsel==1.5.2
python-dateutil==2.8.1
PyYAML==5.3.1
selectorlib==0.16.0
six==1.14.0
w3lib==1.21.0
Werkzeug==1.0.1
